Feat: Add options support
