+++ Cargo.toml
@@ -21,3 +21,4 @@ which = { version = "3.0", default-features = false, features = [] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 lazy_static = "1.0"
+thiserror = "1.0.50"
