+++ src/util/test.rs
@@ -1,0 +1,14 @@
+use super::{find_program, find_version};
+
+#[test]
+fn test_find_program() {
+    if let Some(path) = find_program() {
+        println!("Conan path: {}", path.to_str().unwrap());
+    }
+}
+#[test]
+fn test_find_version() {
+    if let Some(version) = find_version() {
+        println!("Conan version: {}", version);
+    }
+}
